@@ -1,13 +1,5 @@
+import LandingPage from '@/components/marketing/LandingPage'
+
 export default function Home() {
-  return (
-    <div className="min-h-screen bg-white">
-      <h1 className="text-4xl font-bold text-center py-10">🎉 TailwindCSS v4 Successfully Upgraded!</h1>
-      <p className="text-center text-gray-600">Zero-config approach is working perfectly!</p>
-      <div className="text-center mt-4">
-        <span className="inline-block bg-gradient-to-r from-blue-500 to-purple-600 text-white px-4 py-2 rounded-lg">
-          v4.1.11 Ready 🚀
-        </span>
-      </div>
-    </div>
-  );
+  return <LandingPage />;
 }
